requirements.txt
asgiref==3.6.0
Django==4.1.6
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
